+++ report/duct/src/main/java/com/mastercard/test/flow/report/duct/HeadlessGui.java
@@ -1,0 +1,18 @@
+package com.mastercard.test.flow.report.duct;
+
+/**
+ * An empty {@link Gui} for use in headless environments
+ */
+public class HeadlessGui implements Gui {
+
+	@Override
+	public void show() {
+		// nothing to show
+	}
+
+	@Override
+	public void hide() {
+		// nothing to hide
+	}
+
+}
